+++ core/src/components/picker-column-option/test/a11y/picker-column-option.e2e.ts
@@ -10,7 +10,7 @@ configs({ directions: ['ltr'] }).forEach(({ config, title }) => {
     test('should not have accessibility violations', async ({ page }) => {
       await page.goto(`/src/components/picker-column-option/test/a11y`, config);
 
-      const results = await new AxeBuilder({ page }).analyze();
+      const results = await new AxeBuilder({ page }).disableRules('color-contrast').analyze();
 
       expect(results.violations).toEqual([]);
     });